+++ src/main/java/com/example/BlogMovieWebsiteProject/dto/PostsDto.java
@@ -31,7 +31,7 @@ public class PostsDto
     private String description;
     private List<String> tags;
     private List<ItemPostsDto> itemPost;
-    private long yourRating;
+    private double yourRating;
     @Field(type = FieldType.Date, format = DateFormat.custom, pattern = "yyyy-MM-dd HH:mm:ss")
     private Date created;
     private Integer visit;
